sensor reading service now runs in foreground
SensorMonitorService is responsible for reading data from sensors currently.
This may be expanded if needed but as a first pass solution this is ok. The
service acts as a foreground service and will exist even if the application
itself is closed by the user.

Motivation is to ensure that falls can be detected whilst the user is using
another app, or whilst the phones in their pocket, for example.