sensor service now has a separate consume thread
Added a producer-consumer model for sensor readings. Sensor reading events
are handled by a thread which offers the reading to a class which has a private
LinkedBlockingQueue. The consumer thread then blocks on this queue and extracts
the SensorEvent objects and performs data processing. Currently this is just
writing to a debug file. In the future this will probably call some sort
of callback for data analysis and handling.