Added initial fall-like event detection procedure
In short terms, the detection is implemented in a finite state
machine. It first samples the acceleration, performing threshold
checks on the magnitude. If it exceeds this magnitude, it waits for some
time for any further peaks to finish (ie. the body to stop bouncing).
If another peak it detected, it starts over. Then, it waits for the post-fall
/ recovery phase for some time, if another peak happens it goes back to the
start. Some future work will add a test for whether or not what just occured
is a legitimate fall or a false alarm, hopefully afterwards it will be passed
to the FSM.