added medicaid_model to git
Restoring DNNClassifiers from disk is complicated and buggy.
One 'working' method is to leverage the fact that TensorFlow restores the
checkpoint data if any was found, so performing additional training is
possible. We downright abuse this feature to load the DNNClassifier from
disk, as there is no current (working) method to restore the model via the
python API at time of writing.